Run the application in a Docker container
The Dockerfile will use the multi state feature to first build the
application in production mode, then remove all not needed stuff and
finally build a nice and compact image with the compiled assets and serve
it with a minimal http server. This makes running the application
painless, reproducible, portable and secure. See the README.md for
details on how to run it (use the --build flag to rebuild the images
and ignore the cache).