Introduce code editor
So far you had to change the underlying JSON for a warband outside the
app and retrigger a build if you wanted to change something.
By introducing a codemirror editor to the app this can now be done
without recompile.

- Added CodeMirror
- Added option to download JSON file
- Added Toolbar
- Made JSON less verbose by allowing strings next to WeaponReferences
in the weapons array

TODOS:
- CodeMirror cursor positioning when copy&pasting, selecting more than
one line, etc.
- Validate JSON before submitting it to the renderer (errors lead to
an unusable app)
- CodeMirror code completion
- CodeMirror linting
- CodeMirror JSON schema
- Rules and otherEquipment with parameters
- custom army rules (e.g. Orders for Adepta Sororitas)
...
...
...